+++ actix-web-codegen/CHANGES.md
@@ -2,7 +2,7 @@
 
 ## Unreleased - 2021-xx-xx
 * Improve error recovery potential when macro input is invalid. [#2410]
-* Add `#[actix_web::test_rt]` macro for setting up tests with a runtime. [#2409]
+* Add `#[actix_web::test]` macro for setting up tests with a runtime. [#2409]
 * Minimum supported Rust version (MSRV) is now 1.52.
 
 [#2410]: https://github.com/actix/actix-web/pull/2410